Modify app to read database config from database.yml
Note that the config file is a little different than your average rails
app.  First, there's no environments because, well, we're always
connecting (read-only) to the live db.  Second, the fields all have
colons in the file because it was easier than requiring active_support
or writing a method to symbolize the keys.