Helper script to bump "rails"
"rails" isn't a dependency of this app because we only really need
activerecord (and exception-notifyer needs actionmailer).

But, the gems we do need have strict dependencies on each other, which
means you can't update one without updating the other six.

This script adds "rails" as a temporary dependency, updates _that_ to
the most recent version that satisfies the version in `.rails-version`,
then removes the dependency again. Having the end effect of updating
everything "rails" depends on without us having to depend on rails.