Fallback to no specific version if file is missing
Dependapot jsut grabs certain files from the repo to do its analysis.
`.ruby-version` is one of them, `.rails-version` is not. And so,
dependabot currently can't make update PRs, becuase of an exception
while parsing the Gemfile.

This _might_ cause Dependabot to author PRs which update
activerecord-sqlserver-adaptor incorrectly, but that's better than not
making any PRs.